refactor: move out multilayerAxisEntry
