+++ packages/charts/src/chart_types/xy_chart/axes/timeslip/multilayer_ticks.ts
@@ -6,10 +6,15 @@
  * Side Public License, v 1.
  */
 
-import { TimeBin, TimeRaster } from './rasters';
+import { Scale, ScaleContinuous } from '../../../../scales';
+import { XDomain } from '../../domains/types';
+import { Projection } from '../../state/selectors/visible_ticks';
+import { AxisTick } from '../../utils/axis_utils';
+import { rasters, TimeBin, TimeRaster } from './rasters';
 
 const MAX_TIME_TICK_COUNT = 50; // this doesn't do much for narrow charts, but limits tick count to a maximum on wider ones
 const WIDTH_FUDGE = 1.05; // raster bin widths are sometimes approximate, but there's no raster that's just 5% denser/sparser, so it's safe
+const MAX_TIME_GRID_COUNT = 12;
 
 /** @internal */
 export const notTooDense = (
@@ -27,3 +32,84 @@ export const notTooDense = (
     (domainInSeconds * 1000) / maxTickCount <= raster.approxWidthInMs
   );
 };
+
+/** @internal */
+export function multilayerAxisEntry(
+  xDomain: XDomain,
+  extendByOneBin: boolean,
+  range: [number, number],
+  timeAxisLayerCount: any,
+  scale: Scale<string | number> | ScaleContinuous, // fixme it's only the latter for now
+  getMeasuredTicks: (
+    skale: Scale<number | string>,
+    ticks: (number | string)[],
+    layer: number | undefined,
+    detailedLayer: number,
+    labelFormat?: (d: number | string) => string,
+    showGrid?: boolean,
+  ) => Projection,
+) {
+  const rasterSelector = rasters({ minimumTickPixelDistance: 24, locale: 'en-US' }, xDomain.timeZone);
+  const domainValues = xDomain.domain; // todo consider a property or object type rename
+  const domainFromS = Number(domainValues[0]) / 1000; // todo rely on a type guard or check rather than conversion
+  const binWidth = xDomain.minInterval;
+  const domainExtension = extendByOneBin ? binWidth : 0;
+  const domainToS = ((Number(domainValues[domainValues.length - 1]) || NaN) + domainExtension) / 1000;
+  const layers = rasterSelector(notTooDense(domainFromS, domainToS, binWidth, Math.abs(range[1] - range[0])));
+  let layerIndex = -1;
+  const fillLayerTimeslip = (
+    layer: number,
+    detailedLayer: number,
+    timeTicks: number[],
+    labelFormat: (n: number) => string,
+    showGrid: boolean,
+  ) => {
+    return {
+      entry: getMeasuredTicks(
+        scale,
+        timeTicks,
+        layer,
+        detailedLayer,
+        labelFormat as (d: number | string) => string, // todo dissolve assertion
+        showGrid,
+      ),
+      fallbackAskedTickCount: NaN,
+    };
+  };
+  return layers.reduce(
+    (combinedEntry: { ticks: AxisTick[] }, l: TimeRaster<TimeBin>, detailedLayerIndex) => {
+      if (l.labeled) layerIndex++; // we want three (or however many) _labeled_ axis layers; others are useful for minor ticks/gridlines, and for giving coarser structure eg. stronger gridline for every 6th hour of the day
+      if (layerIndex >= timeAxisLayerCount) return combinedEntry;
+      const binWidthS = binWidth / 1000;
+      const { entry } = fillLayerTimeslip(
+        layerIndex,
+        detailedLayerIndex,
+        [...l.binStarts(domainFromS - binWidthS, domainToS + binWidthS)]
+          .filter((b) => b.nextTimePointSec > domainFromS && b.timePointSec <= domainToS)
+          .map((b) => 1000 * b.timePointSec),
+        !l.labeled ? () => '' : layerIndex === timeAxisLayerCount - 1 ? l.detailedLabelFormat : l.minorTickLabelFormat,
+        notTooDense(domainFromS, domainToS, binWidth, Math.abs(range[1] - range[0]), MAX_TIME_GRID_COUNT)(l),
+      );
+      const minLabelGap = 4;
+
+      const lastTick = entry.ticks[entry.ticks.length - 1];
+      if (lastTick && lastTick.position + entry.labelBox.maxLabelBboxWidth > range[1]) {
+        lastTick.axisTickLabel = '';
+      }
+
+      return {
+        ...entry,
+        ...combinedEntry,
+        ticks: (combinedEntry.ticks || []).concat(
+          entry.ticks.filter(
+            (tick, i, a) =>
+              i > 0 ||
+              !a[1] ||
+              a[1].domainClampedPosition - tick.domainClampedPosition >= entry.labelBox.maxLabelBboxWidth + minLabelGap,
+          ),
+        ),
+      };
+    },
+    { ticks: [] }, // this should turn into a full Projection
+  );
+}
